split out user routes
