+++ backend/server.js
@@ -14,7 +14,6 @@ const BankItem = require( './models/BankItem' );
 const Image = require( './models/Image' );
 require('dotenv').config();
 const stripe = require('stripe')(process.env.STRIPE_SECRET_KEY);
-var upload = multer();
 const bcrypt = require('bcrypt');
 const saltRounds = 10;
 
@@ -81,6 +80,7 @@ app.get('/', (req, res, next) => {
 
 app.use('/images/', require("./routes/images"));
 app.use('/donation-request/', require('./routes/donation-request'));
+app.use('/user/', require("./routes/user"));
 
 app.post('/login', function(req, res, next) {
   passport.authenticate('local', function(err, user, info) {
@@ -153,82 +153,6 @@ app.post('/mfaverify', (req, res, next) => {
 		});
 	}
 })
-
-app.get('/user', (req, res, next) => {
-	console.log(req.user);
-	if (req.user) {
-		let loggedInUser = JSON.parse(JSON.stringify(req.user));
-		delete loggedInUser.passwordHash;
-		return res.json(loggedInUser);
-	} else {
-		return res.json({});
-	}
-})
-
-app.post('/UserPage', (req, res, next) => {
-	let updateUser = {};
-	let hash = '';
-	
-	if (req.body.bio) {
-		updateUser.bio = req.body.bio;	
-	} else 
-		updateUser.bio = req.user.bio;
-		
-	if (req.body.first) {
-		updateUser.first = req.body.first;	
-	} else
-		updateUser.first= req.user.first;
-		
-	if (req.body.last) {
-		updateUser.last = req.body.last;	
-	} else 
-		updateUser.last = req.user.last;
-		
-	if (req.body.username) {
-		updateUser.username = req.body.username;	
-	} else 
-		updateUser.username = req.user.username;
-		
-	if (req.body.passwordHash) {
-		bcrypt.hash(req.body.passwordHash, saltRounds, function(err, hash) {
-			updateUser.passwordHash = hash;
-
-			User.updateOne({username: req.user.username}, updateUser).then(updatedUser => {
-				if(updatedUser) {
-					return res.json({Success: 'Successfully updated user', updateUser: updateUser})
-				} else {
-					return res.status(500).json({error: 'Issue updating user'})
-				}
-			})
-		});
-	} else {
-		User.updateOne({username: req.user.username}, updateUser).then(updatedUser => {
-			if(updatedUser) {
-				return res.json({Success: 'Successfully updated user', updateUser: updateUser})
-			} else {
-				return res.status(500).json({error: 'Issue updating user'})
-			}
-		})
-	}
-})
-
-app.post('/user/profilephoto', upload.single('myFile'), (req, res) => {
-	console.log(req.file);
-	Image.create({data: req.file.buffer, name: req.file.originalname, mime: req.file.mimetype}, function(err, image) {
-		if (err) {
-			console.log(err);
-			return res.status(500).json({error: 'Error uploading image. Please try again'});
-		} else {
-			User.updateOne({_id: req.user._id}, {avatar: '/images/' + image._id}).then(response => {
-				if(response) {
-					return res.json({Success: 'Successfully updated profile photo'})
-				} else {
-					return res.status(500).json({error: 'Issue updating profile photo'})
-				}
-			});
-		}
-	});
-});
 
 app.get('/logout', (req, res, next) => {
 	req.logOut();
